% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidict-package.R
\docType{package}
\name{epidict-package}
\alias{epidict}
\alias{epidict-package}
\title{epidict: Epidemiology data dictionaries and random data generators}
\description{
The 'R4EPIs' project \url{https://r4epi.github.io/sitrep} seeks to provide a set of standardized tools for analysis of outbreak and survey data in humanitarian aid settings. This package currently provides standardized data dictionaries from MSF OCA for outbreak scenarios (Acute Jaundice Syndrome, Cholera, Diphtheria, Measles, Meningitis) and surveys (Retrospective mortality and access to care, Malnutrition, Vaccination coverage and Event Based Surveillance). In addition, a data generator from these dictionaries is provided. It is also possible to read in any ODK format data dictionary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R4EPI/epidict}
  \item \url{https://r4epi.github.io/epidict}
}

}
\author{
\strong{Maintainer}: Alexander Spina \email{aspina@appliedepi.org} (\href{https://orcid.org/0000-0001-8425-1867}{ORCID})

Authors:
\itemize{
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})
  \item Lukas Richter
  \item Patrick Keating
}

Other contributors:
\itemize{
  \item Annick Lenglet [contributor]
  \item Applied Epi Incorporated [copyright holder]
  \item Medecins Sans Frontieres Operational Centre Amsterdam [funder]
}

}
\keyword{internal}
