% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf_dict_rename_helper.R
\name{msf_dict_rename_helper}
\alias{msf_dict_rename_helper}
\title{Helper for aligning your data to MSF standardised dictionaries and analysis templates.}
\usage{
msf_dict_rename_helper(dictionary, copy_to_clipboard = TRUE)
}
\arguments{
\item{dictionary}{Specify which MSF dictionary you would like to use.
See \code{\link[=msf_dict]{msf_dict()}} for options.}

\item{copy_to_clipboard}{if \code{TRUE} (default), the rename template will be
copied to the user's clipboard with \code{\link[clipr:write_clip]{clipr::write_clip()}}. If \code{FALSE}, the
rename template will be printed to the user's console.}
}
\value{
A dplyr command used to rename columns in your data frame according
to the dictionary
}
\description{
Helper for aligning your data to MSF standardised dictionaries and analysis templates.
}
