% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf_dict_rename_helper.R
\name{msf_dict_rename_helper}
\alias{msf_dict_rename_helper}
\title{Helper for aligning your data to the dictionary}
\usage{
msf_dict_rename_helper(
  disease,
  varnames = "data_element_shortname",
  copy_to_clipboard = TRUE
)
}
\arguments{
\item{disease}{Specify which disease you would like to use.
Currently supports "Cholera", "Measles", "Meningitis", "AJS",
"Mortality", "Nutrition", "Vaccination_short" and "Vaccination_long".}

\item{varnames}{Specify name of column that contains varnames. Currently
default set to "data_element_shortname".
If \code{dictionary} is a survey ("Mortality", "Nutrition", "Vaccination_short"
or "Vaccination_long") \code{varnames} needs to be "name"`.}

\item{copy_to_clipboard}{if \code{TRUE} (default), the rename template will be
copied to the user's clipboard with \code{\link[clipr:write_clip]{clipr::write_clip()}}. If \code{FALSE}, the
rename template will be printed to the user's console.}
}
\value{
a dplyr command used to rename columns in your data frame according
to the dictionary
}
\description{
Helper for aligning your data to the dictionary
}
